Problem: duplicate lib.rs.cc symbols when used with defi-wallet-core
Close:
- #89
- #94

Solution:
- Add defi-wallet-core-rs as submodule
- Add wrapper crate play-cpp-sdk
