+++ .gitmodules
@@ -4,3 +4,6 @@
 [submodule "demo/include/easywsclient"]
 	path = demo/include/easywsclient
 	url = git@github.com:dhbaird/easywsclient.git
+[submodule "defi-wallet-core-rs"]
+	path = defi-wallet-core-rs
+	url = git@github.com:crypto-com/defi-wallet-core-rs.git
