+++ demo/.env
@@ -8,3 +8,5 @@ export PAY_WEBSOCKET_PORT="4567"
 # Don't share the secret to the clients
 export PAY_WEBHOOK_SIGNATURE_SECRET=""
 export PAY_WEBHOOK_PORT="4567"
+
+source ../defi-wallet-core-rs/scripts/.env
