+++ demo/Makefile
@@ -25,28 +25,48 @@ easywsclient.o.x86_64:
 	arch -x86_64 g++ -c include/easywsclient/easywsclient.cpp
 
 static: easywsclient.o
-	g++ -o cppexamplestatic easywsclient.o main.cc lib/libextra_cpp_bindings.a -std=c++14 $(FLAGS)
-
-dynamic: easywsclient.o
-	g++ -o cppexample \
+	g++ -o demostatic \
 		easywsclient.o \
 		main.cc \
+		chainmain.cc \
+		cronos.cc \
+		extra.cc \
+		lib/libplay_cpp_sdk.a \
+		-std=c++14 $(FLAGS)
+
+dynamic: easywsclient.o
+	g++ -o demo \
+		easywsclient.o \
+		main.cc \
+		chainmain.cc \
+		cronos.cc \
+		extra.cc \
 		include/*.cc \
 		include/extra-cpp-bindings/src/*.cc \
+		include/defi-wallet-core-cpp/src/*.cc \
 		lib/libcxxbridge1.a \
-		-lextra_cpp_bindings \
+		-lplay_cpp_sdk \
 		-std=c++14 \
 		$(FLAGS) \
 		-L lib \
 
 x86_64_build: prepare_x86_64 easywsclient.o.x86_64
-	arch -x86_64 g++ -o cppexample easywsclient.o main.cc lib/libextra_cpp_bindings.a -std=c++14 $(FLAGS)
+	arch -x86_64 \
+		g++ -o demo \
+		easywsclient.o \
+		main.cc \
+		chainmain.cc \
+		cronos.cc \
+		extra.cc \
+		lib/libplay_cpp_sdk.a \
+		-std=c++14 \
+		$(FLAGS)
 
 run_static:
-	./cppexamplestatic
+	./demostatic
 
 run_dynamic:
-	export LD_LIBRARY_PATH=$(PWD)/lib && ./cppexample
+	export LD_LIBRARY_PATH=$(PWD)/lib && ./demo
 
 clean:
-	rm cppexample cppexamplestatic
+	rm demo demostatic
