+++ demo/extra.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void test_blackscout_cronoscan();
+void test_wallet_connect();
