+++ play-cpp-sdk/src/lib.rs
@@ -1,0 +1,2 @@
+pub use defi_wallet_core_cpp::*;
+pub use extra_cpp_bindings::*;
